Add a process above the daemon to restart it
